rename cache to redis
