+++ .ebextensions/00_env_vars.config
@@ -10,7 +10,7 @@ option_settings:
         DB_PASSWD: "{{pull:secretsmanager:/io.cv.app/stg/db:SecretString:password}}"
         DB_USER: "{{pull:secretsmanager:/io.cv.app/stg/db:SecretString:username}}"
         DD_API_KEY: "{{pull:secretsmanager:/io.cv.app/shared/datadog:SecretString:dd_api_key}}"
-        REDIS_URI: "{{pull:secretsmanager:/io.cv.app/stg/cache:SecretString:redis_uri}}"
+        REDIS_URI: "{{pull:secretsmanager:/io.cv.app/stg/redis:SecretString:redis_uri}}"
         DEEP_DAO_BASE_URL: "https://api.deepdao.io"
         MIXPANEL_API_KEY: "{{pull:secretsmanager:/io.cv.app/stg/mixpanel:SecretString:mixpanel_api_key}}"
         NEXT_PUBLIC_TELEGRAM_BOT_ID: "{{pull:secretsmanager:/io.cv.app/shared/telegram:SecretString:telegram_bot_id}}"
