+++ lib/websockets/broadcaster.ts
@@ -2,8 +2,8 @@ import { createAdapter } from '@socket.io/redis-adapter';
 import type { Socket } from 'socket.io';
 import { Server } from 'socket.io';
 
+import { redisClient } from 'adapters/redis/redisClient';
 import { prisma } from 'db';
-import { redisClient } from 'lib/cache/redisClient';
 import { SpaceMembershipRequiredError } from 'lib/permissions/errors';
 
 import type { WebsocketEvent, WebsocketMessage } from './interfaces';
